+++ photutils/detection/daofinder.py
