+++ photutils/detection/irafstarfinder.py
