+++ photutils/detection/peakfinder.py
